Feat: 친구추가 기능 구현

+++ back/src/main/java/com/lemon/backend/domain/friend/service/FriendsService.java
@@ -1,7 +1,7 @@
 package com.lemon.backend.domain.friend.service;
 
 public interface FriendsService {
-    Integer addFriend(Integer userId, Integer friendId);
+    Long addFriend(Integer userId, Integer friendId);
 
     void deleteFriend(Integer userId, Integer friendId);
